Extracted SimulationImpl class
Redmine: #1917

In the future it will become templated and will
be moved out of Simulation hierarchy (i.e. its
instantiation will be included
as a member into Simulation class)
