+++ Core/Simulation/GISASSimulation.h
@@ -15,7 +15,7 @@
 #ifndef GISASSIMULATION_H
 #define GISASSIMULATION_H
 
-#include "Simulation.h"
+#include "SimulationImpl.h"
 
 class MultiLayer;
 class IMultiLayerBuilder;
@@ -25,14 +25,14 @@ class Histogram2D;
 //! Main class to run a Grazing-Incidence Small-Angle Scattering simulation.
 //! @ingroup simulation
 
-class BA_CORE_API_ GISASSimulation : public Simulation
+class BA_CORE_API_ GISASSimulation : public SimulationImpl
 {
 public:
     GISASSimulation();
     GISASSimulation(const MultiLayer& p_sample);
     GISASSimulation(const std::shared_ptr<IMultiLayerBuilder> p_sample_builder);
 
-    ~GISASSimulation() final {}
+    virtual ~GISASSimulation();
 
     GISASSimulation* clone() const { return new GISASSimulation(*this); }
