+++ Core/Simulation/OffSpecSimulation.h
@@ -15,20 +15,20 @@
 #ifndef OFFSPECSIMULATION_H
 #define OFFSPECSIMULATION_H
 
-#include "Simulation.h"
+#include "SimulationImpl.h"
 
 class Histogram2D;
 
 //! Main class to run an off-specular simulation.
 //! @ingroup simulation
 
-class BA_CORE_API_ OffSpecSimulation : public Simulation
+class BA_CORE_API_ OffSpecSimulation : public SimulationImpl
 {
 public:
     OffSpecSimulation();
     OffSpecSimulation(const MultiLayer& p_sample);
     OffSpecSimulation(const std::shared_ptr<class IMultiLayerBuilder> p_sample_builder);
-    ~OffSpecSimulation() final {}
+    virtual ~OffSpecSimulation();
 
     OffSpecSimulation* clone() const { return new OffSpecSimulation(*this); }
